+++ src/modules/quiz-book/reducer.ts
@@ -9,9 +9,6 @@ import {
   SEARCH_QUIZBOOK_LIST,
   SEARCH_QUIZBOOK_LIST_ERROR,
   SEARCH_QUIZBOOK_LIST_SUCCESS,
-  DELETE_QUIZBOOK,
-  DELETE_QUIZBOOK_SUCCESS,
-  DELETE_QUIZBOOK_ERROR,
 } from "./actions";
 import { QuizBookAction, QuizBookState } from "./types";
 import QuizBookModel from "@/common/model/quiz-book";
@@ -84,23 +81,6 @@ const quizBookReducer = createReducer<QuizBookState, QuizBookAction>(
       data: action.payload,
     }),
     [SEARCH_QUIZBOOK_LIST_ERROR]: (state, action) => ({
-      ...state,
-      loading: true,
-      error: action.payload,
-      data: null,
-    }),
-
-    [DELETE_QUIZBOOK]: (state) => ({
-      ...state,
-      loading: true,
-      error: null,
-    }),
-    [DELETE_QUIZBOOK_SUCCESS]: (state, action) => ({
-      ...state,
-      loading: false,
-      data: action.payload,
-    }),
-    [DELETE_QUIZBOOK_ERROR]: (state, action) => ({
       ...state,
       loading: true,
       error: action.payload,
