+++ src/modules/quiz-book/saga.ts
@@ -2,8 +2,6 @@ import quizbookAPI from "@/common/lib/api/quizbook";
 import QuizBookModel from "@/common/model/quiz-book";
 import { call, put, takeEvery } from "redux-saga/effects";
 import {
-  deleteQuizBookAsync,
-  DELETE_QUIZBOOK,
   getQuizBookListAsync,
   GET_QUIZBOOK_LIST,
   postQuizBookLikstAsync,
@@ -11,7 +9,6 @@ import {
   searchQuizBookListAsync,
   SEARCH_QUIZBOOK_LIST,
 } from "./actions";
-import { DeleteResult } from "./types";
 
 function* getQuizBookListSaga(
   action: ReturnType<typeof getQuizBookListAsync.request>
@@ -61,24 +58,8 @@ function* searchQuizBookListSaga(
   }
 }
 
-function* deleteQuizBookSaga(
-  action: ReturnType<typeof deleteQuizBookAsync.request>
-) {
-  try {
-    const { quizBookId } = action.payload;
-    const deleteResult: DeleteResult = yield call(
-      quizbookAPI.deleteQuizBook,
-      quizBookId
-    );
-    yield put(deleteQuizBookAsync.success(deleteResult));
-  } catch (e) {
-    yield put(deleteQuizBookAsync.failure(e));
-  }
-}
-
 export function* quizBookSaga() {
   yield takeEvery(GET_QUIZBOOK_LIST, getQuizBookListSaga);
   yield takeEvery(POST_QUIZBOOK_LIKE, postQuizBookLikeSaga);
   yield takeEvery(SEARCH_QUIZBOOK_LIST, searchQuizBookListSaga);
-  yield takeEvery(DELETE_QUIZBOOK, deleteQuizBookSaga);
 }
