+++ src/modules/user-quizbook/reducer.ts
@@ -3,6 +3,9 @@ import {
   GET_USER_QUIZBOOK,
   GET_USER_QUIZBOOK_SUCCESS,
   GET_USER_QUIZBOOK_ERROR,
+  DELETE_QUIZBOOK,
+  DELETE_QUIZBOOK_SUCCESS,
+  DELETE_QUIZBOOK_ERROR,
 } from "./actions";
 import { UserQuizBookAction, UserQuizBookState } from "./types";
 
@@ -18,7 +21,6 @@ const userQuizBookReducer = createReducer<
 >(initialState, {
   [GET_USER_QUIZBOOK]: (state, isDone) => ({
     ...state,
-    // payload: isDone,
     loading: true,
     error: null,
     data: null,
@@ -29,6 +31,28 @@ const userQuizBookReducer = createReducer<
     data: action.payload,
   }),
   [GET_USER_QUIZBOOK_ERROR]: (state, action) => ({
+    ...state,
+    loading: true,
+    error: action.payload,
+    data: null,
+  }),
+
+  [DELETE_QUIZBOOK]: (state) => ({
+    ...state,
+    loading: true,
+    error: null,
+  }),
+
+  [DELETE_QUIZBOOK_SUCCESS]: (state, action) => ({
+    ...state,
+    loading: false,
+    data: state.data?.filter((quizBook) => {
+      const deletedQuizBookId = action.payload;
+      return quizBook.id !== deletedQuizBookId;
+    }),
+  }),
+
+  [DELETE_QUIZBOOK_ERROR]: (state, action) => ({
     ...state,
     loading: true,
     error: action.payload,
